APIs/i-cue.solutions/v1/python-aiohttp-client/openapi_server/controllers/inventory_management_controller.py
from typing import List, Dict
from aiohttp import web

from openapi_server import util


async def inventory_amazon_ipi_post(request: web.Request, token=None) -> web.Response:
    """Calculate Amazon Inventory Performance Index (IPI)

    Calculate Amazon Inventory Performance Index (IPI)

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)


async def inventory_caryying_cost_post(request: web.Request, token=None) -> web.Response:
    """Carrying Cost

    Carrying Cost

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)


async def inventory_eoq_post(request: web.Request, token=None) -> web.Response:
    """Calculate economic order quantity

    Calculate economic order quantity

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)


async def inventory_fill_rate_post(request: web.Request, token=None) -> web.Response:
    """Calculate fill rate

    Calculate fill rate

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)


async def inventory_financial_imapct_forecast_accuracy_post(request: web.Request, token=None) -> web.Response:
    """Calculate financial impact of forecast accuracy

    Calculate financial impact of forecast accuracy

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)


async def inventory_inventory_turnover_post(request: web.Request, token=None) -> web.Response:
    """Inventroy Turn-over

    Inventroy Turn-over

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)


async def inventory_ltd_post(request: web.Request, token=None) -> web.Response:
    """Calculate lead time demand

    Calculate lead time demand

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)


async def inventory_moq_post(request: web.Request, token=None) -> web.Response:
    """Calculate minimum order quantity

    Calculate minimum order quantity

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)


async def inventory_optimal_service_level_post(request: web.Request, token=None) -> web.Response:
    """Calculate optimal service level

    Calculate optimal service level

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)


async def inventory_reorder_point_post(request: web.Request, token=None) -> web.Response:
    """Re-order Point

    Re-order Point

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)


async def inventory_safety_stock_post(request: web.Request, token=None) -> web.Response:
    """Safety Stock

    Safety Stock

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)


async def inventory_service_level_post(request: web.Request, token=None) -> web.Response:
    """Calculate service level

    Calculate service level

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)


async def inventory_turns_post(request: web.Request, token=None) -> web.Response:
    """Calculate inventory turns

    Calculate inventory turns

    :param token: User Authentication Token
    :type token: str

    """
    return web.Response(status=200)
